+++ matrix_helper.h
@@ -3,8 +3,8 @@
 #include <vector>
 #include "parse.h"
 int get_Matrix_Dim_from_nodes(Element* elementList,int elementListLength);
-void elementList_to_augmented_Matrix( Element* elementList,int elementListLength, std::vector<float>& conductance, std::vector<float>& currents, int matrix_dim);
-void augmented_Matrix_to_definite_matrix(int elementListLength, const std::vector<float> conductance, const std::vector<float> currents, std::vector<float>& conductance_definite, std::vector<float>& currents_definite, int augmented_matrix_dim);
+void elementList_to_augmented_Matrix( Element* elementList,int elementListLength, std::vector<double>& conductance, std::vector<double>& currents, int matrix_dim);
+void augmented_Matrix_to_definite_matrix(int elementListLength, const std::vector<double> conductance, const std::vector<double> currents, std::vector<double>& conductance_definite, std::vector<double>& currents_definite, int augmented_matrix_dim);
 
 #endif
